Update patch set 1
Patch Set 1:

This is just the test. We need the code for the fix as well. Both are in CL 77331.
There is no corresponding fix from master that can be backported, as master rewrote the code that was buggy.

Patch-set: 1
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>