Update patch set 3
Patch Set 3:

This is kinda disgusting. I'd prefer we don't go this route.

How about we filter at a higher layer (only burning the CPU if serving golang.google.cn) and instead parse the HTML and remove the <a> tags based on an href blacklist set of URLs/patterns?

Patch-set: 3
CC: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>