Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > This is kinda disgusting. I'd prefer we don't go this route.
> The first sentence here could have been elided.

Sorry. I didn't mean to imply the contribution was disgusting. We do need to do something, and I'm glad somebody's working on it.

But maintaining files like this that have conditionals all over the place will be disgusting for people in the future to read, IMO.

> > 
> > > 
> > > How about we filter at a higher layer (only burning the CPU if serving golang.google.cn) and instead parse the HTML and remove the <a> tags based on an href blacklist set of URLs/patterns?
> > 
> > I think that's a good idea. I would reduce the scope of this change to replace golang.org/foo with /foo to start, then follow-up with what Brad is suggesting.
> 
> I agree that this is not an elegant approach. But removing all links would bring situations like: "For more details, please check out this post" (while there isn't a link, nor is it clear which post it is). Instead, I would prefer to keep those links and automatically render a tooltip for blacklisted URLs that is shown if you put your mouse over it (a hover overlay, like tensorflow.google.cn, put mouse over the Google logo in the company list to test)

An on-hover overlay of the blacklisted URL sounds fine to me, if China's cool with that. I don't know what the rules are.

Patch-set: 3
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>