Update patch set 4
Patch Set 4: Run-TryBot+1 Code-Review+2

(2 comments)

> Patch Set 4:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2: Run-TryBot+1 Code-Review+1
> > > 
> > > (1 comment)
> > > 
> > > R=go1.11
> > > 
> > > Thanks for bring this up. I think I initially wrote the rules like
> > > 
> > > (OR  x s:(SLLconst [c] y)) && s.Uses == 1 && clobber(s) -> (ORshiftLL  x y [c])
> > > (OR  x (SLLconst [c] y)) -> (ORshiftLL  x y [c])
> > > 
> > > to get combined load rules work. But this somehow gets lost... Would be good to add a test for the combined load optimizations so that it won't get lost again.
> > 
> > Perhaps this is a question for the dev list, but shouldn't the rewrite code be doing this automatically for patterns where there are child nodes that get dropped?
> > 
> > I ran into a similar problem when writing patterns for UBFX/UBFIZ/etc.  When I tried to limit the patterns to only apply when e.g. the child shift/and opcode had only a single user, I ran into problems because previously transformed nodes were laying around keeping the use count artificially high.
> 
> I think that there may be some balance between compilation speed and use count tracking. More automatic use count tracking may require additional checking and maintenance overhead. I don't know if anyone has done some experiments to measure the impact by more automatic use count tracking.

Normally, we don't need to do this -- most of the rules don't care, and we have a dead code elimination pass right after. Avoiding the check and call makes it slightly faster, probably not much, but doesn't hurt. But it seems ok to me for at least experimenting automatic clobbering. What is the criteria for auto clobbering? A value appears on LFS but not RHS?

I suggest that we do this CL in the current form for now. Currently there are not many rules that need the clobbering. If it turns to be more in the future, we can move to auto clobbering.

Patch-set: 4
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2
Label: Run-TryBot=+1