cmd/link: fix up location lists for dsymutil
DNR

dysmutil, used for GOOS=darwin and linkmode=external, doesn't support
base address selection entries in location lists. When running in that
mode, have the linker translate location lists to CU-relative form
instead.

Doing this in the linker is weird, but the compiler doesn't know about
the link mode. The location list format has a ton of addresses in it,
so it's worth avoiding the extra relocations whenever possible.

(A typical location list entry will be 20 bytes, and contain two
addresses. A relocation record is 16 bytes on disk and something like 64
bytes in memory.)

Change-Id: I200246dedaee4f824966f7551ac95f8d7123d3b1