Update patch set 4
Patch Set 4:

There is a version of this page in which I've not defaulted features to "No" if I don't known their state but simply said "Unknown". I can replace it with a " - " instead, in order to make it look nicer.

I don't want to make this about a single editor but so far only @Ramya replied to this and helped move the issue forward.

Thank you for your time and help. And Thank you! @Brad for being as fast as always.

Patch-set: 4
Reviewer: Gerrit User 6473 <6473@62eb7196-b449-3ce5-99f1-c037f21e1705>