cmd/vet: unexported interface fields on %s are ok
For example, the following program is valid:

	type T struct {
		f interface{}
	}

	func main() {
		fmt.Printf("%s", T{"foo"}) // prints {foo}
	}

Since the field is of type interface{}, we might have any value in it.
For example, if we had T{3}, fmt would complain. However, not knowing
what the type under the interface is, we must be conservative.

However, as shown in #17798, we can still issue an error if the field is
not an interface type. In that case, we know for sure if it will be
printable as %s or not. Convert the regression tests added in the CL
fixing that issue to not use interfaces, as was the case in the original
bug report.

Fixes #23563.

Change-Id: I063885955555917c59da000391b603f0d6dce432