Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> R=go1.11
> 
> The toolchain change looks good to me.
> 
> There are a lot of assembly code, with complicated algorithms. I would rather not to have them. The Go implementation is much more readable.
> 
> If fused multiply-add is critical for performance, we should teach the compiler to generate those instructions.

Yes, compiler improvement is more general and valuable than just optimising a specific function and we are planning to add more compiler optimisations for the patterns we find during manual optimisations. But we still want to keep some manual semantic optimisations (not specific to this CL) because it's hard to teach compiler to do the similar things. We'll try to explain the semantic optimisations clearly and briefly.

Patch-set: 1
CC: Gerrit User 16227 <16227@62eb7196-b449-3ce5-99f1-c037f21e1705>