Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (1 comment)

Ian, thanks for bringing this up. I've updated the CL. Now, fields with interface types that satisfy error or fmt.Stringer will still make vet complain. But not those with interface types that don't, such as interface{}.

It's very late in the cycle, and Josh/Aliaskandr haven't chimed in yet, but I do believe this is a regression of sorts - given that the vet rule was added in 1.10 and it will now be ran as part of go test too.

Patch-set: 2
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>