Update patch set 1
Patch Set 1:

The right fix is to change evconst to preserve the type of the input node in this case. The problem is that in evconst:

	case OANDAND_ | CTBOOL_:
		if v.U.(bool) && rv.U.(bool) {
			goto settrue
		}
		goto setfalse

we jump to settrue, which returns an ideal bool, even though the input is non-ideal. There will be many variants of this. Probably in settrue and setfalse, we should check whether the input is has boolean type and if so, use that bool type instead of idealbool.

Patch-set: 1
Reviewer: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>