spec: remove need for separate Function production (cleanup)
The EBNF production

	Function = Signature FunctionBody .

was used in FunctionDecl, MethodDecl, and FunctionLit, but only
for the latter it shortened the syntax slightly.

This change change "inlines" Function which simplifies Function/
MethodDecl and gets rid of the Function production.

This has no impact on the specified language. Also, Function (the
production) is never referred to by the prose, so it's safe to
remove it from the spec.

Finally, neither go/ast or go/parser have a representation of
this production via a corresponding node or parse function, so
no possibly valuable documentation is lost, either.

Change-Id: Ia2875d31c6ec2d2079081ef481e50bad4f43c694