Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> R=1.11
> 
> This may help godoc, but it doesn't really help the source. Now all the Underlying methods which are so symmetric feel irregular. And why does golint care? This seems overreach to me.
> 
> Is this really needed?

I posted the patch for godoc, not for golint - I simply mentioned golint as a plus.

Yes, the source loses a little bit, but I think that's not as important as the godoc, as most people will read the definitions through programs like "go doc" or "godoc" and not the source directly. That's what I tried to explain in the commit message.

Not really a big deal one way or another, but I thought I might as well upload a CL :)

Also, if you decide on not merging, and you think golint is wrong, perhaps we should open an issue for it, like: "don't warn on inconsistent receiver names if they make the source code more readable".

Patch-set: 2
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>