Update patch set 4
Patch Set 4: Code-Review-2

> Patch Set 4:
> 
> > > Patch Set 4:
>  > >
>  > > Andy, can you be the new owner of this page? I haven't see any
>  > activity from Chris or Steve, suggesting that you might be the de
>  > facto new owner.
>  > 
>  > yep yep yep.
>  > 
>  > Florin, give me a bit (1 business day or so) to get up to speed and
>  > then I’ll take a look. If you need it in sooner just let me know.
> 
> Sorry to be _that_ person but... friendly ping :) 
> 
>  > > Patch Set 4:
>  > >
>  > > Andy, can you be the new owner of this page? I haven't see any
>  > activity from Chris or Steve, suggesting that you might be the de
>  > facto new owner.
>  > 
>  > yep yep yep.
>  > 
>  > Florin, give me a bit (1 business day or so) to get up to speed and
>  > then I’ll take a look. If you need it in sooner just let me know.
> 
>  > > Patch Set 4:
>  > >
>  > > Andy, can you be the new owner of this page? I haven't see any
>  > activity from Chris or Steve, suggesting that you might be the de
>  > facto new owner.
>  > 
>  > yep yep yep.
>  > 
>  > Florin, give me a bit (1 business day or so) to get up to speed and
>  > then I’ll take a look. If you need it in sooner just let me know.
> 
> Sorry to be that person but... friendly ping

Hi Florin,
Thanks for being patient and diligent on this CL. I apologize for not getting back to you sooner.

We think it’s best to remove the feature matrix for now. The speed at which features are being added to the various editors/IDEs greatly outpaces the six-month Go release cycle and it is likely that a developer will see out of date information very shortly after a release.

I will send a follow-up CL removing the matrix.

Apologies for the work you put in here only to have it not submitted. Thanks for understanding.

Andy

Patch-set: 4
Reviewer: Gerrit User 22285 <22285@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2