Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> > Patch Set 5:
> > 
> > Hi Steve, any chance I could get another look over this now that I've addressed your comments?
> 
> The merge conflicts are gone but the layout is still vertical. I honestly think it's quite confusing to most people to list the shortcut way and then the longer version of the same command. I would imagine that most people would actually run both as the page is written that way. 
> 
> If they were side by side I think that would convey that you could take either route. 
> I checked this branch today and they are still vertical.

Are you viewing the guide in this change with the corresponding change to the godoc/static CSS in the tools repo that I opened? You need both for them to display side-by-side.

https://go-review.googlesource.com/c/tools/+/76874

I linked to this related change in the Github Issue https://github.com/golang/go/issues/22484 but I didn't think to mention it here as well.

With the tools godoc/static CSS change if the window is greater than 500px wide it will display side-by-side, otherwise it will display vertically. Happy to adjust the width to suit, but in my initial tests 500px wide seemed like a good point to transition for small screens.

Please let me know if you are viewing both changes (css and html) together and still seeing it vertical, and if so which browser you're using.

Patch-set: 5
Reviewer: Gerrit User 21426 <21426@62eb7196-b449-3ce5-99f1-c037f21e1705>