Update patch set 9
Patch Set 9:

> Patch Set 9: -Code-Review
> 
> I'm afraid this is not about synchronization. Some platforms don't return io.EOF when the other side closes the connection but "connection reset by peer". The difference is probably FIN vs RST, but I'm not sure why those stacks behave like that. (Russ? Brad? This is one of those times I get to learn something about TCP.)
> 
> "connection reset by peer" is not a confusing error, so if you want to convert the test to just != io.EOF I wouldn't object.

Makes sense. It can be said that it is sufficient to test only when it is not EOF.
I think that it is not a matter to be strictly involved in this CL regarding the difference in behaviors across platforms.

If the current patchset is ok, please re-run trybot. PTAL.

Patch-set: 9