Update patch set 11
Patch Set 11:

i personally don't have much impression on this change because it looks like this change doesn't fit with the existing way of api error notification. for example, there's already RecordHeaderError for record-protocol error notification. there's no reason to hesitate to add new stuff for reporting errors for other protocols such as the handshake and alert protocols; tls1.3 i-d has a clear view on this kind of things.

i mean, my concern about this sort of change is that it never becomes a roadblock for tls1.3.

Patch-set: 11