Update patch set 1
Patch Set 1:

> > It sounds bad that we used to hide slice grows from the race detector.
> > Is that because we used to use raw memmove when the write barrier was off?
> 
> No, it's because we have a bit of a mismatch in where race annotations are inserted.
> 
> Before this CL, growslice called memmove or typedmemmove. The compiler doesn't race annotate growslice, and memmove/typedmemmove aren't manually annotated in the runtime because usually the compiler annotates calls it generates to them. So none of the layers get race annotations.
> 
> With this CL, growslice calls typedslicecopy. growslice still isn't annotated by the compiler. However, typedslicecopy is manually race-annotated in the runtime because the compiler lowers slice copies before the race annotation pass, so it doesn't get annotated at the call site like all other copy operations do.
> 
> This means that, with this CL, slice grows are now getting race annotations, but they're getting race annotations *everywhere*, including in calls from the runtime.
> 
> I think the "right" solution to this is to fix the inconsistency around typedslicecopy: get the compiler to race-annotate calls to typedslicecopy and growslice at their call sites (the way everything else is done) rather than manually annotating their implementations. Then it can suppress race annotations for them in the runtime, like it does for other things.

I was slightly confused. growslice itself is manually race-annotated in the runtime. It only reports the read of the old slice, while typedslicecopy reports both the read of the old slice and the write of the new slice. This is either an omission from growslice, or it's supposed to be okay because the new slice is fresh memory anyway. I'm not sure. +Dmitry

But I think the solution I proposed is still the right one. typedslicecopy should be annotated by the compiler at the call site like most other operations. (Maybe growslice should be, too.)

Patch-set: 1
CC: Gerrit User 5400 <5400@62eb7196-b449-3ce5-99f1-c037f21e1705>
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>