Update patch set 4
Patch Set 4:

After looking into this some more, I think the root cause is elsewhere, and this CL is working around that failure.

We've constructed an

    OANDAND(OLITERAL(true), OLITERAL(true))

tree where the OLITERALs are "untyped bool" but the OANDAND is "bool". That's wrong.

I'd rather first try to identify and fix the underlying failure. (But if we can't do that, this fix is acceptable.)

Patch-set: 4
Reviewer: Gerrit User 5440 <5440@62eb7196-b449-3ce5-99f1-c037f21e1705>