runtime: reset block, mutex profile counters after rate change
Also, record the last rate change timestamp and use it to fill in
the DurationNano field of the pprof proto message.

There can be subtle race between rate change and sample collection,
which still can cause some samples selected with previous rate
to be sampled. However, we expect that over time the impact of such
samples will become negligible as more samples with the rate
are collected.

Update #23401

Change-Id: I9d32ab56aef11f58b0351fa0dc5a0fdefbacf0c1