Update patch set 1
Patch Set 1:

This could likely be handled by mid stack inlining too. However making it explicit avoids relying on a heuristic for inlining and avoids calculation that the function should be inlined. It is a small line for line change. I am not sure if mid stack inlining is guaranteed to be always on for go1.11 and this will bring nice savings for such a small change.

It would be nice if there was a way to link (export) a function like this to multiple packages inside the std lib but the original name can be kept. Then we could deduplicate it with newobject.

Patch-set: 1
Reviewer: Gerrit User 16006 <16006@62eb7196-b449-3ce5-99f1-c037f21e1705>