cmd/asm, cmd/internal/obj/ppc64: avoid unnecessary load zeros
The instructions and, or, xor, add allow constant operands,
but in many cases the assembler was generating 2 instructions
to generate the constant value when it could have been done
with 1 instruction.

This change adds more cases to the optab to identify the cases
where the operands are small enough to be loaded with a single
instruction.

Example of previous code:
	oris r3, r0, 0
	ori  r3, r3, 65533

now:
	ori r3, r0, 65533

Updates #23845

Change-Id: I7fdf2320c95815d99f2755ba77d0c6921cd7fad7