cmd/vet: use type information in isLocalType
Now that vet always has type information, there's no reason to use
string handling on type names to gather information about them, such as
whether or not they are a local type.

The new isLocalType is, in intent, almost equivalent to the old one. The
only difference is that it doesn't have special logic for *_test package
names. The old code was added without tests for that however, and I
cannot see how those lines would help - they concern the package path,
not the package name.

Add proper test package versus non-test package test cases with local
types, to make sure that the check does work with the new version of
isLocalType.

Change-Id: I71386b4196922e4c9f2653d90abc382efbf01b3c