Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 2:
> 
> (3 comments)
> 
> I wonder if it would be worth in this CL adding a new ppc64leenc.s file to cmd/asm/internal/asm/testdata to start to fix #18072 on ppc64le? It would be nice to verify that each form of the changed instructions is encoded correctly.

Are you suggesting I just add the file with only these new instructions? Or the complete set of opcodes and operand combinations?

I agree we should have a ppc64enc.s file here and was not aware of the issue you reference. My preference would be to create a separate CL with a more complete set of optab combinations with their encodings, but that would take some work and I'd rather not hold this one up for that. And at the same time review and update all the stale comments in asm9.go.

Patch-set: 2
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>