Update patch set 4
Patch Set 4:

> Patch Set 2:
> 
> (2 comments)
> 
> > Patch Set 2:
> > 
> > (3 comments)
> > 
> > I wonder if it would be worth in this CL adding a new ppc64leenc.s file to cmd/asm/internal/asm/testdata to start to fix #18072 on ppc64le? It would be nice to verify that each form of the changed instructions is encoded correctly.
> 
> Are you suggesting I just add the file with only these new instructions? Or the complete set of opcodes and operand combinations?

I was thinking just the changed ones for now and then at some later time it can be updated to be more complete (or just added to as further changes are made). That way we can be sure that the instructions this CL touches are assembled correctly.

BTW does this have any impact on binary size? Might be worth adding to the commit message if it does.

Patch-set: 4
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>